+++ demo/src/main/java/com/example/test/Main.java
@@ -4,14 +4,15 @@ import java.io.IOException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.example.test.FileProcessing.FileProcessor;
+import com.example.test.service.FileProcessor;
+import com.example.test.utils.MainArgParser;
 
 public class Main {
 
     public static void main(String[] args) throws IOException {
         
         Logger logger = LoggerFactory.getLogger(Main.class);
-        MainConfig config = new MainConfig(args);
+        MainArgParser config = new MainArgParser(args);
 
         FileProcessor fProcessor = new FileProcessor(config);
         try{
