+++ demo/src/main/java/com/example/test/service/JsonWriter.java
@@ -1,4 +1,4 @@
-package com.example.test.FileProcessing;
+package com.example.test.service;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
@@ -11,6 +11,7 @@ import java.time.LocalDate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.example.test.utils.LocalDateAdapter;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
