+++ demo/src/main/java/com/example/test/service/XHTMLParser.java
@@ -1,4 +1,4 @@
-package com.example.test.FileProcessing;
+package com.example.test.service;
 
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
@@ -7,7 +7,7 @@ import org.jsoup.select.Elements;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.example.test.Product.Product;
+import com.example.test.bean.Product;
 
 import java.io.File;
 import java.io.IOException;
