+++ demo/src/main/java/com/example/test/utils/MainArgParser.java
@@ -1,4 +1,4 @@
-package com.example.test;
+package com.example.test.utils;
 
 import org.apache.commons.cli.BasicParser;
 import org.apache.commons.cli.CommandLine;
@@ -10,11 +10,13 @@ import org.apache.commons.cli.ParseException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.example.test.Main;
+
 import lombok.Getter;
 
 @Getter
-public class MainConfig {
-    
+public class MainArgParser {
+
     private Boolean isFolderInput;
     private String inputFolder;
     private String outputFolder;
@@ -26,7 +28,7 @@ public class MainConfig {
     private HelpFormatter formatter = new HelpFormatter();
     private Options options = new Options();
 
-    public MainConfig(String[] args){
+    public MainArgParser(String[] args){
         
         Logger logger = LoggerFactory.getLogger(Main.class);
