+++ demo/src/test/java/ProductGroupTest.java
@@ -5,10 +5,10 @@ import java.util.Map;
 
 import org.junit.Test;
 
-import com.example.test.Product.GroupedProduct;
-import com.example.test.Product.Product;
-import com.example.test.Product.ProductGrouper;
-import com.example.test.Currency.Currency;
+import com.example.test.bean.Currency;
+import com.example.test.bean.GroupedProduct;
+import com.example.test.bean.Product;
+import com.example.test.service.ProductGrouper;
 
 import org.junit.Assert;
